Remove <double> casts from scancode and getstate
Since it wasn't really necessary to do so in any way.

I also patched one crash for when getbf, getgf, or getopp is used on the lua end. This way,

+++ src/inp/AsyncKB.cpp
@@ -21,11 +21,11 @@ bool hasEvent() {
     return core_hasEvent();
 }
 
-double getScanCode() {
+int getScanCode() {
     return core_getScanCode();
 }
 
-double getState() {
+int getState() {
     return core_getState();
 }
